Adds functionality to Notifications
Adds functions for creating a new chore or reward approval
notiications. They can be called with Notification.new_chore
or Notification.new_reward and they each take the id of the
reward or chore that needs approval and a parent id, which will
usually be current_parent.id, but you could pass in a different
id if you wanted to.